requirements.txt
Django>=3.2
numpy
scipy
scikit-image
scikit-learn
Pillow
matplotlib
